test: add get uuid from endpoint util test
